refactor(xapiri): extract costs tab functions → tab_costs.go
Mechanical move per ADR 0014. No behaviour change.
